+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Change import of `exportToCsv` from `stripes-util` to `stripes-components`. Refs UIIN-3025.
 * ECS: Disable opening item details if a user is not affiliated with item's member tenant. Fixes UIIN-3187.
 * Correctly depend on `inflected`. Refs UIIN-3203.
+* Decrease the amount of rerenders in `ConsortialHoldings` component. Fixes UIIN-3196.
 * Sort holdings by location name and call number. Refs UIIN-3207.
 
 ## [12.0.12](https://github.com/folio-org/ui-inventory/tree/v12.0.12) (2025-01-27)