+++ CHANGELOG.md
@@ -15,6 +15,9 @@
 * ECS: Disable opening item details if a user is not affiliated with item's member tenant. Fixes UIIN-3187.
 * Correctly depend on `inflected`. Refs UIIN-3203.
 * Decrease the amount of rerenders in `ConsortialHoldings` component. Fixes UIIN-3196.
+* Sort holdings by location name and call number. Refs UIIN-3207.
+* Remove hover-over text next to "Shelving order" on the Item record detail view. Refs UIIN-3210.
+* CI: Switch to centralized/shared workflow from https://github.com/folio-org/.github. Fixes UIIN-3218.
 * Add Version history button and Version history pane to details view of Item. Refs UIIN-3172.
 
 ## [12.0.12](https://github.com/folio-org/ui-inventory/tree/v12.0.12) (2025-01-27)